fix(aq1.6): QA-2 — split runtime.rs (RULE-003), displaced-flag live test, AC1 assertion
- Extract the daemon-lease lifecycle (GraftReceiverLeaseClient trait,
  RegisteredGraftReceiver, LeaseRefreshBackoff, tick_lease_refresh,
  run_bounded_lease_call) out of runtime.rs into runtime/lease_client.rs.
  No behavior change: public/crate paths preserved via a `pub(crate) use`
  re-export of GraftReceiverLeaseClient (the only symbol crate::lib
  needs), the rsh002 backoff unit test moves into lease_client's own
  #[cfg(test)] mod alongside the type it exercises, and the previously
  private `listener` field on RegisteredGraftReceiver becomes
  `pub(super)` so the parent loop and its test module keep direct access.
  atm-graft-python and hermes-atm depend on none of this (verified: they
  only import GraftClient/HostNudge/HostNudgeInjector/MailboxWorkCounts),
  so neither needs any change.
- Add ATM-QA-102: a live-loop test where a refresh tick observes NotOwner
  from a foreign generation that took over the lease while this receiver
  was already announced — proves `displaced` gets set, refresh becomes a
  permanent no-op (no further refresh calls, success or failure), and the
  accept loop keeps delivering, all through the real running receive loop.
- Add AC1: a dedicated test asserting the registered lease's endpoint,
  capability, and owner generation literally equal the bound listener's
  own values (team/agent via the registry's lookup key), closing the gap
  where the only prior coverage checked registration counts and
  generation inequality against a stale lease (AC5), never equality
  against the bind inputs.

Verified: cargo test --workspace --all-targets (1256 passed), cargo
clippy --workspace --all-targets -D warnings, cargo fmt --all --check,
python3 .just/run_lint.py (32/32, including `lines` RULE-003 and
sc-boundary), scripts/test_atm_graft_python.py.
